pyproject.toml
[project]
name = "pdf-pal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie>=1.6.1",
    "chromadb>=1.5.2",
    "groq>=1.0.0",
    "loguru>=0.7.3",
    "pathlib>=1.0.1",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.5",
    "reportlab>=4.4.10",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
